+++ service-product/module-core/src/main/java/com/github/msafriends/serviceproduct/modulecore/dto/productimage/ProductImageResponse.java
@@ -1,4 +1,4 @@
-package com.github.msafriends.serviceproduct.moduleapi.dto;
+package com.github.msafriends.serviceproduct.modulecore.dto.productimage;
 
 import com.github.msafriends.serviceproduct.modulecore.domain.productimage.ProductImage;
